requirements.txt
fastmcp
fastapi
uvicorn[standard]
httpx
langchain-community
langchain-core
langchain-ollama
pydantic